+++ src/components/IconButton/story.tsx
@@ -1,0 +1,21 @@
+import { boolean, select, text } from '@storybook/addon-knobs/react';
+import { storiesOf } from '@storybook/react';
+import React from 'react';
+import { PropsType } from '.';
+import IconButton, { PropsType as PlainPropsType } from '../IconButton';
+import { MediumIcons } from '../Icon/types';
+
+/* tslint:disable */
+const mediumIconKeys = Object.keys(MediumIcons).filter(key => MediumIcons[key as any].match('<svg'));
+/* tslint:enable */
+
+storiesOf('IconButton', module).add('Default', () => {
+    return (
+        <IconButton
+            icon={select('icon', mediumIconKeys, 'cart') as PropsType['icon']}
+            loading={boolean('loading', false)}
+            variant={select('variant', ['default', 'destructive'], 'default') as PlainPropsType['variant']}
+            title={text('title', 'Click me')}
+        />
+    );
+});
