+++ src/components/Modal/test.tsx
@@ -6,9 +6,9 @@ import BreakpointProvider from '../BreakpointProvider';
 import { PropsType } from '../BreakpointProvider/';
 import StyledModal, { StyledModalWrapper } from './style';
 import TransitionAnimation from '../TransitionAnimation';
-import { IconButton } from '../Button';
+import IconButton from '../IconButton';
 
-jest.mock('../ScrollBox', () => jest.fn().mockImplementation((_: PropsType): string => 'div'));
+jest.mock('../ScrollBox', () => jest.fn().mockImplementation((): string => 'div'));
 
 jest.mock('../BreakpointProvider', () => {
     return jest.fn().mockImplementation((props: PropsType): JSX.Element => {
